Refactor: Comprehensive codebase cleanup and health improvements
This commit addresses a broad range of issues identified during a full codebase review, focusing on removing duplicates, eliminating legacy/mock code, improving clarity, and addressing potential bugs.

Key changes include:

1.  **Consolidated Utility Functions:**
    *   Merged `src/services/api/conversationUtils.ts` and `src/services/utils/conversationUtils.ts`, making the latter the single source of truth.
    *   Centralized `getSystemPrompt` logic, removing duplications from `chatApi.ts` and `aiService.ts`.

2.  **Removed Legacy/Mock Code:**
    *   Deleted mock `src/services/authService.ts`; Supabase auth via `AuthContext` is used.
    *   Deleted legacy `src/services/ai/providerService.ts` (mock AI provider).
    *   Deleted `src/services/api/messageCreators.ts` as its functions were redundant to `messageService.sendMessage`.

3.  **Refactored for Clarity and Consistency:**
    *   Removed `src/contexts/AuthContext.tsx` re-export file, updating imports to point to source files in `src/contexts/auth/`.
    *   Relocated shadcn/ui toast implementation to `src/components/ui/toast/internalToastImplementation.ts` for better co-location and clarity.

4.  **Addressed Potential Bugs/Improvements:**
    *   Adjusted `TOAST_REMOVE_DELAY` in toast implementation to 5 seconds (from 1000 seconds).
    *   Noted that `generateDocument` and `generateInterviewQuestions` in `src/services/aiService.ts` are currently mock implementations and are candidates for future backend integration.

5.  **General Code Health:**
    *   Verified that deleted files are no longer imported.
    *   These changes improve maintainability, reduce confusion, and streamline the codebase.

Recommendations for you:
*   Run linters and code formatters.
*   Execute any existing test suites to ensure no regressions.
