+++ src/components/chat/ModernChatInterface.tsx
@@ -1,7 +1,7 @@
 
 import { useState, useEffect, useRef } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "@/contexts/AuthContext";
+import { useAuth } from "@/contexts/auth/useAuth";
 import { toast } from "sonner";
 import { conversationService, Conversation, Message } from "@/services/conversationService";
 import MessagesList from "./components/MessagesList";
