+++ src/components/chat/ModernChatSidebar.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 import { Separator } from "@/components/ui/separator";
 import { toast } from "sonner";
 import { conversationService, Conversation, ConversationType } from "@/services/conversationService";
-import { useAuth } from "@/contexts/AuthContext";
+import { useAuth } from "@/contexts/auth/useAuth";
 import { useIsMobile } from "@/hooks/use-mobile";
 import SidebarHeader from "./sidebar/SidebarHeader";
 import SidebarGroups from "./sidebar/SidebarGroups";
