+++ src/components/linkedin/LinkedInOptimizer.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { LinkedInProfile } from "@/contexts/AuthContext";
+import { LinkedInProfile } from "@/contexts/auth/types";
 import { 
   Award, 
   CheckCircle,
