--- src/contexts/AuthContext.tsx
@@ -1,17 +1,0 @@
-
-import { AuthProvider, AuthContext } from './auth/AuthContext';
-import { useAuth } from './auth/useAuth';
-import { AuthProviderType, Profile, LinkedInProfile, AuthContextType } from './auth/types';
-
-export {
-  AuthProvider,
-  AuthContext,
-  useAuth
-};
-
-export type {
-  AuthProviderType,
-  Profile,
-  LinkedInProfile,
-  AuthContextType
-};
