+++ src/services/api/chatApi.ts
@@ -1,5 +1,6 @@
 
 import { ConversationMessage, ConversationType } from "../aiService";
+import { getSystemPrompt } from "../utils/conversationUtils";
 
 interface ChatRequestBody {
   messages: {
@@ -28,7 +29,7 @@ export const generateChatResponse = async (
     const formattedMessages = [
       {
         role: 'system',
-        content: getSystemPromptForType(conversationType)
+        content: getSystemPrompt(conversationType)
       },
       ...history.map(msg => ({
         role: msg.role as 'user' | 'assistant',
@@ -63,17 +64,3 @@ export const generateChatResponse = async (
     throw error;
   }
 };
-
-// Helper function to get system prompts based on conversation type
-function getSystemPromptForType(type: ConversationType): string {
-  switch(type) {
-    case 'resume':
-      return "You are a professional resume writer and career coach. Help the user create or improve their resume. Provide specific, tailored advice based on their experience and the job they're targeting. Be concise but thorough, and format content in a clean, professional way.";
-    case 'cover_letter':
-      return "You are an expert at writing compelling cover letters. Help the user create a cover letter that highlights their relevant experience and skills for the specific job they're applying to. Be professional, authentic, and persuasive.";
-    case 'interview_prep':
-      return "You are an interview coach with expertise in preparing candidates. Help the user prepare for job interviews by providing common questions, strategies for effective answers, and feedback on their practice responses. Be supportive but honest in your assessment.";
-    default:
-      return "You are a helpful career development assistant. Provide guidance, advice, and support for various career-related questions and needs.";
-  }
-}
